requirements.txt
requests
python-dotenv
Pillow
discord.py
transformers
pytz
discord-py-interactions
interactions-wait-for
